COMP: Fix dashboard warnings. Mostly by adding local variable initializations. Also, add explicit casts to avoid signed/unsigned comparison warnings.

+++ Code/Review/d1mach.c
@@ -23,6 +23,6 @@ double d1mach_(int *i)
     case 4: return DBL_EPSILON;
     case 5: return log10((double)FLT_RADIX); 
     } 
-  fprintf(stderr, "invalid argument: d1mach(%ld)\n", *i); 
+  fprintf(stderr, "invalid argument: d1mach(%ld)\n", (long) *i);
   return 0; /* some compilers demand return values */
 }
